+++ workspaces/marketplace/plugins/marketplace-backend/src/router.ts
@@ -16,7 +16,7 @@
 
 import express, { Request, Response, NextFunction } from 'express';
 import Router from 'express-promise-router';
-import { NotAllowedError } from '@backstage/errors';
+import { InputError, NotAllowedError } from '@backstage/errors';
 import {
   HttpAuthService,
   PermissionsService,
@@ -44,6 +44,7 @@ import { removeVerboseSpecContent } from './utils/removeVerboseSpecContent';
 import { rules as extensionRules } from './permissions/rules';
 import { matches } from './utils/permissionUtils';
 import { InstallationDataService } from './installation/InstallationDataService';
+import { ConfigFormatError } from './errors/ConfigFormatError';
 
 export type MarketplaceRouterOptions = {
   httpAuth: HttpAuthService;
@@ -176,6 +177,39 @@ export async function createRouter(
     },
   );
 
+  router.post(
+    '/package/:namespace/:name/configuration',
+    requireInitializedInstallationDataService,
+    async (req, res) => {
+      const marketplacePackage = await marketplaceApi.getPackageByName(
+        req.params.namespace,
+        req.params.name,
+      );
+      if (!marketplacePackage.spec?.dynamicArtifact) {
+        throw new Error(
+          `Package ${marketplacePackage.metadata.name} is missing 'spec.dynamicArtifact'`,
+        );
+      }
+
+      const newConfig = req.body.configYaml;
+      if (!newConfig) {
+        throw new InputError("'configYaml' object must be present");
+      }
+      try {
+        installationDataService.updatePackageConfig(
+          marketplacePackage.spec.dynamicArtifact,
+          newConfig,
+        );
+      } catch (e) {
+        if (e instanceof ConfigFormatError) {
+          throw new InputError(e.message);
+        }
+        throw e;
+      }
+      res.status(200).json({ status: 'OK' });
+    },
+  );
+
   router.get('/plugins', async (req, res) => {
     const request = decodeGetEntitiesRequest(createSearchParams(req));
     const plugins = await marketplaceApi.getPlugins(request);
@@ -282,36 +316,52 @@ export async function createRouter(
     },
   );
 
-  router.post('/plugin/:namespace/:name/configuration', async (req, res) => {
-    // installs the plugin
-    const installDecision = await authorizeConditional(
-      req,
-      extensionsPluginWritePermission,
-    );
-    if (installDecision.result === AuthorizeResult.DENY) {
-      throw new NotAllowedError(
-        `Not allowed to configure ${req.params.namespace}:${req.params.name}`,
+  router.post(
+    '/plugin/:namespace/:name/configuration',
+    requireInitializedInstallationDataService,
+    async (req, res) => {
+      // installs the plugin
+      const installDecision = await authorizeConditional(
+        req,
+        extensionsPluginWritePermission,
       );
-    }
+      if (installDecision.result === AuthorizeResult.DENY) {
+        throw new NotAllowedError(
+          `Not allowed to configure ${req.params.namespace}:${req.params.name}`,
+        );
+      }
 
-    const plugin = await marketplaceApi.getPluginByName(
-      req.params.namespace,
-      req.params.name,
-    );
-
-    const hasInstallAccess =
-      installDecision.result === AuthorizeResult.ALLOW ||
-      (installDecision.result === AuthorizeResult.CONDITIONAL &&
-        matches(plugin, installDecision.conditions));
-
-    if (!hasInstallAccess) {
-      throw new NotAllowedError(
-        `Not allowed to configure ${req.params.namespace}:${req.params.name}`,
+      const plugin = await marketplaceApi.getPluginByName(
+        req.params.namespace,
+        req.params.name,
       );
-    }
 
-    res.status(200).json({});
-  });
+      const hasInstallAccess =
+        installDecision.result === AuthorizeResult.ALLOW ||
+        (installDecision.result === AuthorizeResult.CONDITIONAL &&
+          matches(plugin, installDecision.conditions));
+
+      if (!hasInstallAccess) {
+        throw new NotAllowedError(
+          `Not allowed to configure ${req.params.namespace}:${req.params.name}`,
+        );
+      }
+
+      const newConfig = req.body.configYaml;
+      if (!newConfig) {
+        throw new InputError("'configYaml' object must be present");
+      }
+      try {
+        await installationDataService.updatePluginConfig(plugin, newConfig);
+      } catch (e) {
+        if (e instanceof ConfigFormatError) {
+          throw new InputError(e.message);
+        }
+        throw e;
+      }
+      res.status(200).json({ status: 'OK' });
+    },
+  );
 
   router.get('/plugin/:namespace/:name/packages', async (req, res) => {
     const packages = await marketplaceApi.getPluginPackages(