+++ workspaces/orchestrator/plugins/orchestrator/src/components/WorkflowInstancePage/WorkflowInstancePageContent.tsx
@@ -16,42 +16,32 @@
 
 import React, { useState } from 'react';
 import { useAsync } from 'react-use';
-import useObservable from 'react-use/esm/useObservable';
 
-import {
-  CodeSnippet,
-  Content,
-  InfoCard,
-  Link,
-} from '@backstage/core-components';
-import { appThemeApiRef, useApi } from '@backstage/core-plugin-api';
+import { Content, InfoCard, Link } from '@backstage/core-components';
+import { useApi } from '@backstage/core-plugin-api';
 import { usePermission } from '@backstage/plugin-permission-react';
 
-import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
 import Grid from '@mui/material/Grid';
-import { useTheme } from '@mui/material/styles';
 import Typography from '@mui/material/Typography';
 import moment from 'moment';
 import { makeStyles } from 'tss-react/mui';
 
 import {
   AssessedProcessInstanceDTO,
-  capitalize,
   InputSchemaResponseDTO,
   orchestratorAdminViewPermission,
   ProcessInstanceDTO,
   WorkflowDataDTO,
 } from '@red-hat-developer-hub/backstage-plugin-orchestrator-common';
 
-import { orchestratorApiRef } from '../../src/api/api';
-import { VALUE_UNAVAILABLE } from '../constants';
-import { InfoDialog } from './InfoDialog';
+import { orchestratorApiRef } from '../../api/api';
+import { VALUE_UNAVAILABLE } from '../../constants';
+import { WorkflowRunDetail } from '../WorkflowRunDetail';
+import { WorkflowRunDetails } from '../WorkflowRunDetails';
+import { VariablesDialog } from './VariablesDialog';
 import { WorkflowInputs } from './WorkflowInputs';
 import { WorkflowProgress } from './WorkflowProgress';
 import { WorkflowResult } from './WorkflowResult';
-import { WorkflowRunDetail } from './WorkflowRunDetail';
-import { WorkflowRunDetails } from './WorkflowRunDetails';
 
 export const mapProcessInstanceToDetails = (
   instance: ProcessInstanceDTO,
@@ -93,48 +83,6 @@ const useStyles = makeStyles()(() => ({
   },
   recommendedLabel: { margin: '0 0.25rem' },
 }));
-
-const VariablesDialogContent = ({
-  instanceVariables,
-}: {
-  instanceVariables: WorkflowDataDTO;
-}) => {
-  const appThemeApi = useApi(appThemeApiRef);
-  const activeThemeId = useObservable(
-    appThemeApi.activeThemeId$(),
-    appThemeApi.getActiveThemeId(),
-  );
-  const theme = useTheme();
-
-  return (
-    <Box>
-      {Object.entries(instanceVariables).map(([key, value]) => (
-        <div key={key} style={{ marginBottom: '16px' }}>
-          <Typography variant="h6" style={{ marginBottom: '8px' }}>
-            {capitalize(key)}
-          </Typography>
-          <CodeSnippet
-            text={JSON.stringify(value, null, 2)}
-            language="json"
-            showLineNumbers
-            showCopyCodeButton
-            customStyle={{
-              color:
-                activeThemeId === 'dark'
-                  ? theme.palette.grey[100]
-                  : theme.palette.grey[800],
-              backgroundColor:
-                activeThemeId === 'dark'
-                  ? theme.palette.grey[900]
-                  : theme.palette.grey[100],
-              padding: '25px 0',
-            }}
-          />
-        </div>
-      ))}
-    </Box>
-  );
-};
 
 export const WorkflowInstancePageContent: React.FC<{
   assessedInstance: AssessedProcessInstanceDTO;
@@ -202,23 +150,11 @@ export const WorkflowInstancePageContent: React.FC<{
 
   return (
     <Content noPadding>
-      <InfoDialog
-        title="Run Variables"
-        onClose={toggleVariablesDialog}
+      <VariablesDialog
         open={isVariablesDialogOpen}
-        dialogActions={
-          <Button
-            color="primary"
-            variant="contained"
-            onClick={toggleVariablesDialog}
-          >
-            Close
-          </Button>
-        }
-        wideDialog
-      >
-        <VariablesDialogContent instanceVariables={instanceVariables} />
-      </InfoDialog>
+        onClose={toggleVariablesDialog}
+        instanceVariables={instanceVariables}
+      />
       <Grid container spacing={2}>
         <Grid item xs={6}>
           <InfoCard