+++ workspaces/adoption-insights/plugins/adoption-insights/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @red-hat-developer-hub/backstage-plugin-adoption-insights
 
+## 0.2.0
+
+### Minor Changes
+
+- 3a80c30: Backstage version bump to v1.39.1
+
+### Patch Changes
+
+- 3f93232: Updated dependency `@mui/icons-material` to `5.17.1`.
+  Updated dependency `@mui/material` to `5.17.1`.
+  Updated dependency `@red-hat-developer-hub/backstage-plugin-theme` to `^0.8.0`.
+- Updated dependencies [3a80c30]
+  - @red-hat-developer-hub/backstage-plugin-adoption-insights-common@0.3.0
+
 ## 0.1.0
 
 ### Minor Changes