+++ workspaces/marketplace/plugins/catalog-backend-module-marketplace/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @red-hat-developer-hub/backstage-plugin-catalog-backend-module-marketplace
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [7ae0ca0]
+  - @red-hat-developer-hub/backstage-plugin-marketplace-common@0.5.0
+
 ## 0.3.1
 
 ### Patch Changes